third_party/WebKit/LayoutTests/hittesting/image-with-clip-path-expected.txt

Image hit testing should not include clipped-out regions.

 
PASS successfullyParsed is true

TEST COMPLETE
PASS document.elementFromPoint(75, 125).id is not 'imageWithReferenceClipPath'
PASS document.elementFromPoint(125, 75).id is 'imageWithReferenceClipPath'
PASS document.elementFromPoint(75+100, 125).id is not 'imageWithShapeClipPath'
PASS document.elementFromPoint(125+100, 75).id is 'imageWithShapeClipPath'
PASS document.elementFromPoint(100, 225).id is not 'transformedWithReferenceClipPath'
PASS document.elementFromPoint(135, 175).id is not 'transformedWithReferenceClipPath'
PASS document.elementFromPoint(115, 175).id is 'transformedWithReferenceClipPath'
PASS document.elementFromPoint(120, 225).id is 'transformedWithReferenceClipPath'
PASS document.elementFromPoint(100+100, 225).id is not 'transformedWithShapeClipPath'
PASS document.elementFromPoint(135+100, 175).id is not 'transformedWithShapeClipPath'
PASS document.elementFromPoint(115+100, 175).id is 'transformedWithShapeClipPath'
PASS document.elementFromPoint(120+100, 225).id is 'transformedWithShapeClipPath'
